python-backend/requirements.txt
# 基础依赖
fastapi==0.104.1
uvicorn==0.36.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# 可选依赖 - 高级功能
aiofiles==23.2.1
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.2.1
httpx==0.25.2
jsonschema==4.20.0
croniter==2.0.1

# AutoGen依赖 (可选)
autogen-agentchat
autogen-ext[openai]